5 - Initial home screen design
The bottom nav bar doesn't yet work. However its still a good reference
for the what the UI will look like. It's tricky becuase we will need to
be able to change the state of the nav bar from other places other than
the 3 buttons on it. For instance tapping a popular places card should
navigate to the make request screen. It's not quite the same as named
routes like we previously used. So will be left as a task for later.
Since we likely need to setup a global bloc to handle its state.